Further refactoring to unify sync and async implementations.
The old Http class has been split up. We have now:

* HttpCore implements the core primitives to do Ably-ready requests, ie.
  HTTP requests on the current thread that implement Ably auth, know how
  to reauthorize, and interpret Ably's error responses as ErrorInfos.
  It also has ancillary inner classes like Response, BodyHandler. 
* HttpScheduler, that implements operations on top of a HttpCore
  scheduled to an Executor. The API is asynchronous, but the underlying
  Executor may make its behaviour synchronous. We have both
  SyncHttpScheduler and AsyncHttpScheduler subclasses.
* Http, which offers a higher level API to SyncHttpScheduler and 
  AsyncHttpScheduler to help expose a synchronous or asynchronous API
  out of HTTP operations without duplicating code.
* HttpHelpers, for helper methods used primarily in tests. It's also
  used in some non-test places that don't need the sync/async
  unification so I didn't bother converting them to use Http, which
  would make them more complex.
* HttpConstants to hold constants like method names, header names, etc.

PaginatedQuery/Result and AsyncPaginatedQuery/Result have now unified
implementations too. Now both wrap BasePaginatedQuery/Result. To make
this user-friendly some internal bridging classes boilerplate had to be
added, unfortunately, but the end result should be simple enough.

Tests have not been converted yet.
