--- lib/src/main/java/io/ably/lib/http/SyncHttp.java
@@ -1,14 +1,0 @@
-package io.ably.lib.http;
-
-import io.ably.lib.util.CurrentThreadExecutor;
-
-/**
- * Created by tcard on 31/8/17.
- */
-
-public class SyncHttp extends CallbackfulHttp<CurrentThreadExecutor> {
-
-    public SyncHttp(Http http) {
-        super(http, CurrentThreadExecutor.INSTANCE);
-    }
-}
