+++ lib/src/main/java/io/ably/lib/http/SyncHttpScheduler.java
@@ -1,0 +1,13 @@
+package io.ably.lib.http;
+
+import io.ably.lib.util.CurrentThreadExecutor;
+
+/**
+ * A HttpScheduler that runs everything in the current thread.
+ */
+public class SyncHttpScheduler extends HttpScheduler<CurrentThreadExecutor> {
+
+    public SyncHttpScheduler(HttpCore httpCore) {
+        super(httpCore, CurrentThreadExecutor.INSTANCE);
+    }
+}
